Reorganize the OpenAI platform files
